Remove unintended double computation
Fix untyped floating point literals that were causing unintended
float-to-double promotion.
Also replace expensive libm calls with cheaper (*f) alternatives
to avoid similar promotion.
Update qsort reference output (slightly less precise).

+++ src/susan_corners/susan.c
@@ -321,7 +321,7 @@ void setup_brightness_lut(uint8_t **bp, int32_t thresh, int32_t form)
         temp = temp * temp;
         if (form == 6)
             temp = temp * temp * temp;
-        temp = 100.0 * exp(-temp);
+        temp = 100.0f * expf(-temp);
         *(*bp + k) = (uint8_t)temp;
     }
 }
